+++ guake/world_map_layout.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8; -*-
+"""
+Data management for the World Map layout.
+Handles loading, saving, and modifying the world_map.json file.
+"""
